pyproject.toml
[tool.poetry]
name = "movie-genie"
version = "0.1.0"
description = "LLM-based recommender that suggests great movies for you and your friends"
authors = ["Ignacio Avas <gmail@ignacioavas.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kaggle = "^1.6.14"
pandas = "^2.2.2"
httpx = "^0.27.0"
pyarrow = "^16.1.0"
tqdm = "^4.66.4"
pillow = "^10.3.0"
weaviate-client = "^4.6.4"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.0"
orjson = "3.9.6"
fastapi = "^0.111.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
